webapi/tct-filesystem-tizen-tests/filesystem/File_copyTo_dir_overwrite_false.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Samsung Electronics Co., Ltd.

Licensed under the Apache License, Version 2.0 (the License);
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.


Authors:
        Mariusz Polasinski <m.polasinski@samsung.com>

-->

<html>

<head>
<title>File_copyTo_dir_overwrite_false</title>
<script type="text/javascript" src="../resources/testharness.js"></script>
<script type="text/javascript" src="../resources/testharnessreport.js"></script>
<script type="text/javascript" src="support/filesystem_common.js"></script>
</head>

<body>
<div id="log"></div>
<script type="text/javascript">
//==== TEST: File_copyTo_dir_overwrite_false
//==== LABEL Check if error callback is invoked when copy a directory to another location where directory of that name already exists (overwrite is false)
//==== PRIORITY P2
//==== SPEC Tizen Web API:IO:Filesystem:File:copyTo M
//==== SPEC_URL https://developer.tizen.org/help/topic/org.tizen.web.device.apireference/tizen/filesystem.html
//==== TEST_CRITERIA MERRCB


var resolveSuccess, resolveError, copySuccess, copyError, fsTestDir1, fsTestDirName1, fsTestSubDir1, fsTestSubDirName1,
    fsTestDir2, fsTestDirName2, fsTestSubDir2, expected = "IOError",
    t = async_test(document.title);

t.step(function () {
    fsTestDirName1 = getDirName("filesystem1");
    fsTestSubDirName1 = getDirName("filesystem1sub");
    fsTestDirName2 = getDirName("filesystem2");

    copySuccess = t.step_func(function () {
        assert_unreached("copySuccess callback shouldn't be invoked");
    });

    copyError = t.step_func(function (error) {
        assert_equals(error.name, expected, "wrong error type");
        t.done();
    });

    resolveSuccess = t.step_func(function (dir) {
        fsTestDir1 = dir.createDirectory(fsTestDirName1);
        fsTestSubDir1 = fsTestDir1.createDirectory(fsTestSubDirName1);
        fsTestDir2 = dir.createDirectory(fsTestDirName2);
        fsTestSubDir2 = fsTestDir2.createDirectory(fsTestSubDirName1);
        fsTestDir1.copyTo(fsTestSubDir1.fullPath, fsTestSubDir2.path, false, copySuccess, copyError);
    });

    resolveError = t.step_func(function (error) {
        assert_unreached("resolve() error callback invoked: name:" + error.name + "msg:" + error.message);
    });

    prepareForTesting(t, function () {
        tizen.filesystem.resolve("documents", resolveSuccess, resolveError, "rw");
    });
});

</script>
</body>
</html>
